requirements.txt
flask
line-bot-sdk
wheel
grpcio
google-cloud-firestore
flask_cors
line-bot-sdk
google-cloud-storage
pandas
gunicorn
google-cloud-logging
tensorflow-cpu == 2.12.0
Pillow